requirements.txt
coverage==5.2.1
Django>=3.2
mock==4.0.2
djangorestframework
django-picklefield==3.0.1
